requirements.txt
ansible-rulebook
pytest-asyncio

